requirements.txt
jupyter
python-dotenv
python-telegram-bot
yfinance
